Change move and merge architecture to use reducers
