+++ package.json
@@ -20,9 +20,6 @@
     ]
   },
   "main": "mobile/index.tsx",
-  "jest": {
-    "resetMocks": false
-  },
   "scripts": {
     "start:expo": "expo start",
     "start": "react-scripts start",
