+++ src/components/Thought.tsx
@@ -27,7 +27,7 @@ import {
   appendToPath,
   equalArrays,
   equalPath,
-  hashContext,
+  hashPath,
   head,
   headId,
   headValue,
@@ -171,7 +171,7 @@ const mapStateToProps = (state: State, props: ThoughtContainerProps) => {
   const isCursorGrandparent =
     !isExpandedHoverTopPath && !!cursor && equalPath(rootedParentOf(state, parentOf(cursor)), path)
 
-  const isExpanded = !!expanded[hashContext(path)]
+  const isExpanded = !!expanded[hashPath(path)]
   const isLeaf = !hasChildren(state, contextLive)
 
   return {
