+++ src/reducers/index.ts
@@ -44,7 +44,7 @@ export { default as invalidState } from './invalidState'
 export { default as isPushing } from './isPushing'
 export { default as join } from './join'
 export { default as loadLocalState } from './loadLocalState'
-export { default as mergePending } from './mergePending'
+export { default as mergeThoughts } from './mergeThoughts'
 export { default as modalComplete } from './modalComplete'
 export { default as moveThought } from './moveThought'
 export { default as moveThoughtDown } from './moveThoughtDown'
