+++ src/reducers/moveThought.ts
@@ -1,12 +1,12 @@
 import _ from 'lodash'
 import { treeMove } from '../util/recentlyEditedTree'
-import { rerank, updateThoughts } from '../reducers'
+import { rerank, updateThoughts, mergeThoughts } from '../reducers'
 import { rootedParentOf, getThoughtById, getChildrenRankedById } from '../selectors'
 import { Path, SimplePath, State } from '../@types'
 
 // util
-import { head, isDescendantPath, reducerFlow } from '../util'
-import move from '../selectors/move'
+import { appendToPath, head, isDescendantPath, normalizeThought, pathToContext, reducerFlow, timestamp } from '../util'
+import { getSessionId } from '../util/sessionManager'
 
 export interface MoveThoughtPayload {
   oldPath: Path
@@ -41,11 +41,76 @@ const moveThought = (state: State, { oldPath, newPath, offset, skipRerank, newRa
 
   const newCursorPath = isPathInCursor ? (isCursorAtOldPath ? newPath : getUpdateCursor()) : state.cursor
 
-  const { thoughtIndexUpdates, contextIndexUpdates, pendingMerges } = move(state, {
-    sourceThoughtPath: oldPath,
-    destinationThoughtPath: rootedParentOf(state, newPath),
-    rank: newRank,
-  })
+  const sourceThoughtPath = oldPath
+  const destinationThoughtPath = rootedParentOf(state, newPath)
+  const sourceThoughtId = head(sourceThoughtPath)
+  const destinationThoughtId = head(destinationThoughtPath)
+
+  const sourceThought = getThoughtById(state, sourceThoughtId)
+  const sourceParentThought = getThoughtById(state, sourceThought.parentId)
+  const destinationThought = getThoughtById(state, destinationThoughtId)
+
+  const sameContext = sourceParentThought.id === destinationThoughtId
+  const childrenOfDestination = getChildrenRankedById(state, destinationThoughtId)
+
+  /**
+   * Find duplicate thought.
+   */
+  const duplicateSubthought = () =>
+    childrenOfDestination.find(child => normalizeThought(child.value) === normalizeThought(sourceThought.value))
+
+  // if thought is being moved to the same context that is not a duplicate case
+  const duplicateThought = !sameContext ? duplicateSubthought() : undefined
+
+  const isPendingMerge = duplicateThought && (sourceThought.pending || duplicateThought.pending)
+
+  const newStateAfterMerge =
+    !!duplicateThought &&
+    !isPendingMerge &&
+    mergeThoughts(state, {
+      sourceThoughtPath,
+      targetThoughtPath: appendToPath(destinationThoughtPath, duplicateThought.id),
+    })
+
+  const destinationContext = pathToContext(state, destinationThoughtPath)
+
+  const isArchived = destinationContext?.indexOf('=archive') !== -1
+
+  // if move is used for archive then update the archived field to latest timestamp
+  const archived = isArchived ? timestamp() : destinationThought.archived
+
+  const contextIndexUpdates = newStateAfterMerge
+    ? {}
+    : {
+        // if moved within same context we don't need to remove and add source thought from source parent to the destination
+        ...(!sameContext
+          ? {
+              // remove source thought from the previous source parent children array
+              [sourceParentThought.id]: {
+                ...sourceParentThought,
+                children: sourceParentThought.children.filter(thoughtId => thoughtId !== sourceThought.id),
+                lastUpdated: timestamp(),
+                updatedBy: getSessionId(),
+              },
+              // add source thought to the destination thought children array
+              [destinationThought.id]: {
+                ...destinationThought,
+                children: [...destinationThought.children, sourceThought.id],
+                lastUpdated: timestamp(),
+                updatedBy: getSessionId(),
+              },
+            }
+          : {}),
+        // update source thought parent id, rank and other stuffs
+        [sourceThought.id]: {
+          ...sourceThought,
+          parentId: destinationThought.id,
+          rank: newRank,
+          archived,
+          lastUpdated: timestamp(),
+          updatedBy: getSessionId(),
+        },
+      }
 
   return reducerFlow([
     state => ({
@@ -53,14 +118,23 @@ const moveThought = (state: State, { oldPath, newPath, offset, skipRerank, newRa
       cursor: newCursorPath,
       ...(offset != null ? { cursorOffset: offset } : null),
     }),
-
     // update thoughts
-    updateThoughts({
-      contextIndexUpdates,
-      thoughtIndexUpdates,
-      recentlyEdited,
-      pendingMerges,
-    }),
+    newStateAfterMerge
+      ? null
+      : updateThoughts({
+          contextIndexUpdates,
+          thoughtIndexUpdates: {},
+          recentlyEdited,
+          pendingMerges:
+            isPendingMerge && duplicateThought
+              ? [
+                  {
+                    sourcePath: appendToPath(destinationThoughtPath, sourceThought.id),
+                    targetPath: appendToPath(destinationThoughtPath, duplicateThought.id),
+                  },
+                ]
+              : [],
+        }),
 
     // rerank context if ranks are too close
     // skip if this moveThought originated from a rerank
@@ -78,7 +152,7 @@ const moveThought = (state: State, { oldPath, newPath, offset, skipRerank, newRa
           return ranksTooClose ? rerank(state, rootedParentOf(state, newPath) as SimplePath) : state
         }
       : null,
-  ])(state)
+  ])(newStateAfterMerge || state)
 }
 
 export default _.curryRight(moveThought, 2)
