+++ src/reducers/setCursor.ts
@@ -8,7 +8,7 @@ import {
   TUTORIAL_STEP_AUTOEXPAND_EXPAND,
 } from '../constants'
 import { chain, expandThoughts, getSetting, getAllChildren, simplifyPath } from '../selectors'
-import { equalPath, hashContext, headValue, isDescendant, pathToContext } from '../util'
+import { equalPath, hashPath, headValue, isDescendant, pathToContext } from '../util'
 import { settings } from '../reducers'
 import { Index, Path, SimplePath, State, TutorialChoice } from '../@types'
 import globals from '../globals'
@@ -96,7 +96,7 @@ const setCursor = (
    */
   const hasThoughtCollapsed = () =>
     state.cursor &&
-    !expanded[hashContext(state.cursor)] &&
+    !expanded[hashPath(state.cursor)] &&
     (getAllChildren(state, context).length > 0 ||
       (state.cursor.length > (thoughtsResolved || []).length &&
         thoughtsResolved &&
