+++ src/selectors/index.ts
@@ -14,7 +14,6 @@ export { default as getAncestorBy, getAncestorByValue } from './getAncestorBy'
 export { default as getChildPath } from './getChildPath'
 export { default as getContexts } from './getContexts'
 export { default as getContextForThought } from './getContextForThought'
-export { default as getPathForThought } from './getPathForThought'
 export { default as getContextsSortedAndRanked } from './getContextsSortedAndRanked'
 export { getDescendantContexts, getDescendantThoughtIds, someDescendants } from './getDescendants'
 export { default as getEditingPath } from './getEditingPath'
