+++ src/util/index.ts
@@ -33,6 +33,7 @@ export { getQueryParam } from './getQueryParam'
 export { getUserRef } from './getUserRef'
 export { getThoughtIdByContext } from './getThoughtIdByContext'
 export { hashContext } from './hashContext'
+export { hashPath } from './hashPath'
 export { hashThought } from './hashThought'
 export { head } from './head'
 export { headId } from './headId'
